Renamed patches module to bezier
